tools/requirements_lock.txt
# Copyright Offene Werkstatt Wädenswil
# SPDX-License-Identifier: MIT
#
# Python dependencies for maco_gateway and integration tests
# Generated manually - update versions as needed

# ASCON lightweight crypto (for gateway AEAD encryption)
ascon==0.0.9

# Async HTTP client/server (for MockHttpServer and FirebaseClient)
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
multidict==6.0.5
yarl==1.9.4

# AES-128 for NTAG key diversification (used by simulator console)
pycryptodome==3.21.0

# Common dependencies
idna==3.10
